ref(tour): Remove  isAvailable state
