+++ static/app/components/tours/components.spec.tsx
@@ -26,39 +26,10 @@ describe('Tour Components', () => {
   });
 
   describe('TourContextProvider', () => {
-    it('renders children regardless of availability', () => {
-      mockUseTourReducer.mockReturnValue(emptyTourContext);
-
-      const {container: availableContainer} = render(
-        <TourContextProvider<TestTour>
-          isAvailable
-          isCompleted={false}
-          orderedStepIds={ORDERED_TEST_TOUR}
-          tourContext={TestTourContext}
-        >
-          <div>Child Content</div>
-        </TourContextProvider>
-      );
-      expect(within(availableContainer).getByText('Child Content')).toBeInTheDocument();
-
-      const {container: unavailableContainer} = render(
-        <TourContextProvider<TestTour>
-          isAvailable={false}
-          isCompleted={false}
-          orderedStepIds={ORDERED_TEST_TOUR}
-          tourContext={TestTourContext}
-        >
-          <div>Child Content</div>
-        </TourContextProvider>
-      );
-      expect(within(unavailableContainer).getByText('Child Content')).toBeInTheDocument();
-    });
-
     it('renders children regardless of completion', () => {
       mockUseTourReducer.mockReturnValue(emptyTourContext);
       render(
         <TourContextProvider<TestTour>
-          isAvailable
           isCompleted
           orderedStepIds={ORDERED_TEST_TOUR}
           tourContext={TestTourContext}
@@ -79,7 +50,6 @@ describe('Tour Components', () => {
         <TourContextProvider<TestTour>
           isCompleted={false}
           orderedStepIds={ORDERED_TEST_TOUR}
-          isAvailable
           tourContext={TestTourContext}
         >
           <div>Child Content</div>
@@ -91,7 +61,6 @@ describe('Tour Components', () => {
         <TourContextProvider<TestTour>
           isCompleted={false}
           orderedStepIds={ORDERED_TEST_TOUR}
-          isAvailable
           tourContext={TestTourContext}
           omitBlur
         >
@@ -120,7 +89,6 @@ describe('Tour Components', () => {
           tourKey={tourKey}
           isCompleted={false}
           orderedStepIds={ORDERED_TEST_TOUR}
-          isAvailable
           tourContext={TestTourContext}
         >
           <div>Child Content</div>
@@ -140,7 +108,6 @@ describe('Tour Components', () => {
       mockUseTourReducer.mockReturnValue(emptyTourContext);
       const {unmount: unmountInactive} = render(
         <TourContextProvider<TestTour>
-          isAvailable
           isCompleted={false}
           orderedStepIds={ORDERED_TEST_TOUR}
           tourContext={TestTourContext}
@@ -166,7 +133,6 @@ describe('Tour Components', () => {
       });
       render(
         <TourContextProvider<TestTour>
-          isAvailable
           isCompleted={false}
           orderedStepIds={ORDERED_TEST_TOUR}
           tourContext={TestTourContext}
@@ -194,7 +160,6 @@ describe('Tour Components', () => {
       });
       const {unmount: unmountFirstStep} = render(
         <TourContextProvider<TestTour>
-          isAvailable
           isCompleted={false}
           orderedStepIds={ORDERED_TEST_TOUR}
           tourContext={TestTourContext}
@@ -229,7 +194,6 @@ describe('Tour Components', () => {
 
       const {unmount: unmountSecondStep} = render(
         <TourContextProvider<TestTour>
-          isAvailable
           isCompleted={false}
           orderedStepIds={ORDERED_TEST_TOUR}
           tourContext={TestTourContext}
@@ -260,7 +224,6 @@ describe('Tour Components', () => {
 
       render(
         <TourContextProvider<TestTour>
-          isAvailable
           isCompleted={false}
           orderedStepIds={ORDERED_TEST_TOUR}
           tourContext={TestTourContext}
@@ -290,7 +253,6 @@ describe('Tour Components', () => {
       });
       render(
         <TourContextProvider<TestTour>
-          isAvailable
           isCompleted={false}
           orderedStepIds={ORDERED_TEST_TOUR}
           tourContext={TestTourContext}
@@ -342,7 +304,6 @@ describe('Tour Components', () => {
       render(
         <TourContextProvider<TestTour>
           tourKey={tourKey}
-          isAvailable
           isCompleted={false}
           orderedStepIds={ORDERED_TEST_TOUR}
           tourContext={TestTourContext}
